requirements.txt
# Core dependencies
openai>=1.0.0
langchain-core>=0.1.0
langgraph>=0.0.1

# Google Drive integration
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0

# PDF processing
pdf2image>=1.16.0
Pillow>=10.0.0

# Document generation
python-docx>=1.0.0

# Web framework for dashboard
fastapi>=0.100.0
uvicorn>=0.22.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Deep Agents framework (if available, otherwise implement custom)
# deepagents>=0.1.0
